[update]nyartoolkit +NyARDoubleMatrix34にsetValue関数(2)を追加 +SingleMarkerProcessorのsetConfidenceThreshold関数の引数ミスを修正 +
git-svn-id: https://svn.sourceforge.jp/svnroot/nyartoolkit/NyARToolkit/trunk@415 7cac0a50-4618-4814-88d0-24b83990f816

+++ src/jp/nyatla/nyartoolkit/processor/SingleARMarkerProcesser.java
@@ -303,7 +303,7 @@ public abstract class SingleARMarkerProcesser
 	 * @param i_new_detect_cf
 	 * @param i_exist_detect_cf
 	 */
-	public void setConfidenceThreshold(int i_new_cf,double i_exist_cf)
+	public void setConfidenceThreshold(double i_new_cf,double i_exist_cf)
 	{
 		this._detectmarker_cb.cf_threshold_exist=i_exist_cf;
 		this._detectmarker_cb.cf_threshold_new=i_new_cf;
